0.FrontEnd/dogwalk_frontend/requirements.txt
requests==2.32.5
streamlit==1.51.0
mysql-connector-python==9.0.0
bcrypt==5.0.0